[xy] Add database manager.
